+++ src/Events/TranslationHasBeenSet.php
@@ -4,27 +4,13 @@ namespace Spatie\Translatable\Events;
 
 class TranslationHasBeenSet
 {
-    /** @var \Spatie\Translatable\Translatable */
-    public $model;
-
-    /** @var string */
-    public $key;
-
-    /** @var string */
-    public $locale;
-
-    public $oldValue;
-    public $newValue;
-
-    public function __construct($model, string $key, string $locale, $oldValue, $newValue)
-    {
-        $this->model = $model;
-
-        $this->key = $key;
-
-        $this->locale = $locale;
-
-        $this->oldValue = $oldValue;
-        $this->newValue = $newValue;
+    public function __construct(
+        public mixed $model,
+        public string $key,
+        public string $locale,
+        public mixed $oldValue,
+        public mixed $newValue,
+    ) {
+        //
     }
 }