+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to `laravel-translatable` will be documented in this file
 
+## 5.0.0 - unreleased
+
+- require PHP 8+
+- convert syntax to PHP 8
+- drop support for PHP 7.x
+- drop support for Laravel 6.x
+- implement `spatie/laravel-package-tools`
+
 ## 4.6.0 - 2020-11-19
 
 - add support for PHP 8.0 (#241)